requirements.txt
tinkoff-investments==0.2.0b107
grpcio==1.67.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0
pytz==2024.2
tabulate==0.9.0
python-dotenv==1.0.1
